Import tickets - There is nothing I can do but to approve it, because more fixes on Import Tickets are now done on a different branch than "importTickets"....

+++ app/src/main/java/com/wallet/crypto/alphawallet/viewmodel/ConfirmationViewModel.java
@@ -115,12 +115,12 @@ public class ConfirmationViewModel extends BaseViewModel {
 
         if (sends != null && sends.size() > 0)
         {
-            short[] ticketIDs = new short[sends.size()];
+            int[] ticketIDs = new int[sends.size()];
             int index = 0;
 
             for (Integer i : sends)
             {
-                ticketIDs[index++] = i.shortValue();
+                ticketIDs[index++] = i;
             }
 
             int firstIndex = iDs.get(0);
